add type conversion, improve error handling

+++ routes/shifts.js
@@ -39,16 +39,12 @@ router
         location.type = "Point";
       }
 
-      if (
-        !name ||
-        !programId ||
-        !startTime ||
-        !endTime ||
-        !location ||
-        !volunteerLimit
-      ) {
+      if (!name || !programId || !startTime || !endTime || !location) {
         throw error(400, "Insufficient data");
       }
+
+      startTime = Date(startTime);
+      endTime = Date(endTime);
 
       const existingShift = await Shift.findOne({
         name,
@@ -56,6 +52,7 @@ router
         startTime,
         endTime,
         location,
+        volunteerLimit,
       });
 
       if (existingShift) {
@@ -100,6 +97,11 @@ router
   .patch(async (req, res, next) => {
     try {
       const shiftId = req.params.id;
+
+      if (!mongoose.Types.ObjectId.isValid(shiftId)) {
+        throw error(400, "Invalid shift ID");
+      }
+
       const body = req.body;
 
       if (!body || !Object.keys(body).length) {
@@ -125,6 +127,11 @@ router
   .delete(async (req, res, next) => {
     try {
       const shiftId = req.params.id;
+
+      if (!mongoose.Types.ObjectId.isValid(shiftId)) {
+        throw error(400, "Invalid shift ID");
+      }
+
       const result = await Shift.findByIdAndDelete(shiftId);
 
       if (result) {
@@ -147,15 +154,16 @@ router
         throw error(400, "Invalid shift ID");
       }
 
-      const { startTime, endTime } = req.query;
-      const query = { shiftId };
+      const role = req.query.role;
 
-      if (startTime) {
-        query.startTime = { $gte: startTime };
+      if (role && role.some((role) => !allowedRoles.includes(role))) {
+        throw error(400, "Invalid role");
       }
 
-      if (endTime) {
-        query.endTime = { $lte: endTime };
+      let query = { shiftId };
+
+      if (role) {
+        query.role = role;
       }
 
       const result = await ShiftAssignment.find(query)
@@ -165,7 +173,7 @@ router
       if (result) {
         res.json({ shiftAssignment: result });
       } else {
-        throw error(400, "Bad request");
+        throw error(404, "Shift not found");
       }
     } catch (err) {
       next(err);
@@ -202,7 +210,7 @@ router
           }).populate("userId"),
         });
       } else {
-        throw error(400, "Bad request");
+        throw error(404, "Shift or user not found");
       }
     } catch (err) {
       next(err);
@@ -214,7 +222,7 @@ router.delete("/:id/users/:userId/?", async (req, res, next) => {
     const shiftId = req.params.id;
 
     if (!mongoose.Types.ObjectId.isValid(shiftId)) {
-      throw error(400, "Invalid user ID");
+      throw error(400, "Invalid shift ID");
     }
 
     const userId = req.params.userId;
